j
Now explain my panel system, toast popups, and how I want the users navigation through the game to be also mention we are using Godot4.4
also I want to eventually add a character dialous system with a story that has branching paths. keep this in mind for future proofing my game. events will be handled on a calenday system and we need a system that will help me place characters around the overworld depending on the day. perfereable an other CSV system for all this info. I want all the events and dialouge to be pushed this way. .

Also keep in ming. the goal is to have not many scripts if we can d things into other scripts that make sense then we must do so. like have a script just for 1 popup makes no sense by the end of this all. I should only need to mess with my CSV files to add tot he game, other than of course placing and creating tiles and assets.